Log number of cells found in block for given app.
