+++ src/rpc.rs
@@ -224,7 +224,7 @@ pub async fn get_cells(
 ) -> Result<Vec<Option<Vec<u8>>>, String> {
 	let begin = SystemTime::now();
 
-	let store_size = (msg.max_rows * msg.max_cols) as usize;
+	let store_size = (msg.dimensions.rows * msg.dimensions.cols) as usize;
 	let store: Arc<Mutex<Vec<Option<Vec<u8>>>>> = Arc::new(Mutex::new(vec![None; store_size]));
 
 	let store_0 = store.clone();
@@ -234,10 +234,10 @@ pub async fn get_cells(
 
 	stream::iter(cells_and_store)
 		.for_each_concurrent(num_cpus::get(), |(row, col, store)| async move {
-			let proof = get_kate_query_proof_by_cell(url, msg.num, row as u16, col as u16).await;
+			let proof = get_kate_query_proof_by_cell(url, msg.number, row as u16, col as u16).await;
 
 			let mut handle = store.lock().unwrap();
-			handle[col * msg.max_rows as usize + row] = match proof {
+			handle[col * msg.dimensions.rows + row] = match proof {
 				Ok(v) => Some(v),
 				Err(e) => {
 					log::info!("error: {}", e);
@@ -249,8 +249,8 @@ pub async fn get_cells(
 
 	log::info!(
 		"Received {} cells of block {}\t{:?}",
-		msg.max_cols * msg.max_rows,
-		msg.num,
+		msg.dimensions.cols * msg.dimensions.rows,
+		msg.number,
 		begin.elapsed().unwrap()
 	);
